Extract common superclass "TimingSessionContainer"
