+++ src/test/java/net/obvj/performetrics/util/print/PrintUtilsTest.java
@@ -45,6 +45,7 @@ import net.obvj.performetrics.Counter;
 import net.obvj.performetrics.Counter.Type;
 import net.obvj.performetrics.Performetrics;
 import net.obvj.performetrics.Stopwatch;
+import net.obvj.performetrics.TimingSessionContainer;
 import net.obvj.performetrics.config.ConfigurationHolder;
 import net.obvj.performetrics.util.DurationFormat;
 
@@ -69,7 +70,7 @@ class PrintUtilsTest
         Performetrics.setDefaultPrintStyle(DEFAULT_PRINT_STYLE);
     }
 
-    Stopwatch stopwatch = mock(Stopwatch.class);
+    TimingSessionContainer stopwatch = mock(Stopwatch.class);
     PrintStyle printStyle = mock(PrintStyle.class);
 
     @BeforeEach
